+++ client/src/test/java/org/realityforge/gwt/appcache/client/TestApplicationCache.java
@@ -19,8 +19,9 @@ final class TestApplicationCache
   }
 
   @Override
-  public void swapCache()
+  public boolean swapCache()
   {
+    return true;
   }
 
   @Override